New feature: check_password and check_password_caps now can run RemotePostBuildEvent if TLPI_POSTBUILD_RUN_ROOT=1 .

+++ TLPI/ch08-users_groups/check_password.cpp
@@ -52,7 +52,7 @@ main(int argc, char *argv[])
 
 	printf("Username: ");
 	fflush(stdout);
-	if (fgets(username, lnmax, stdin) == NULL)
+	if (fgets(username, (int)lnmax, stdin) == NULL)
 		exit(EXIT_FAILURE);             /* Exit on EOF */
 
 	len = strlen(username);
